# src/V1.0.0/docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Recourse API",
    "contact": {},
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
    "/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/logIn"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/logIn"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/logIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "LogIn Successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Email or password not match",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              }
            }
          }
        }
      }
    },
    "/v1/resource": {
      "post": {
        "tags": ["Resource"],
        "summary": "Create Resource",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createResource"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/createResource"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/createResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "LogIn Successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Email or password not match",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Resource"],
        "summary": "Get Resource list",
        "parameters": [
          {
            "name": "searchText",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a resource",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token provided not found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              }
            }
          },
          "403": {
            "description": "Failed to authentication",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse403"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse403"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse403"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              }
            }
          }
        }
      }
    },
    "/v1/resource/{id}": {
      "get": {
        "tags": ["Resource"],
        "summary": "Get Resource by Id",
        "parameters": [
          {
            "name": "id",
            "description": "id of resource",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a recourse",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token provided not found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              }
            }
          },
          "403": {
            "description": "Failed to authentication",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse403"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse403"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse403"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Resource"],
        "summary": "Update Resource by Id",
        "parameters": [
          {
            "name": "id",
            "description": "id of resource",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createResource"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/createResource"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/createResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token provided not found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              }
            }
          },
          "403": {
            "description": "Failed to authentication",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse403"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse403"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse403"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Resource"],
        "summary": "Delete Resource by Id",
        "parameters": [
          {
            "name": "id",
            "description": "id of resource",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token provided not found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse401"
                }
              }
            }
          },
          "403": {
            "description": "Failed to authentication",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse403"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse403"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse403"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse500"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "logIn": {
        "required": ["email", "password"],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Enter the email of user"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Enter the Password of user"
          }
        },
        "example": {
          "email": "test@yopmail.com",
          "password": "TestPassword@123"
        },
        "additionalProperties": false
      },
      "createResource": {
        "required": ["title", "description"],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of resource"
          },
          "description": {
            "type": "string",
            "description": "Description of resource"
          }
        },
        "example": {
          "title": "title",
          "description": "description"
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "message"
          },
          "code": {
            "type": "integer",
            "example": "200"
          },
          "data": {
            "type": "object"
          }
        },
        "additionalProperties": false
      },
      "errorResponse400": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "message"
          },
          "code": {
            "type": "integer",
            "example": "400"
          },
          "data": {
            "type": "object"
          }
        },
        "additionalProperties": false
      },
      "errorResponse401": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "message"
          },
          "code": {
            "type": "integer",
            "example": "401"
          },
          "data": {
            "type": "object"
          }
        },
        "additionalProperties": false
      },
      "errorResponse403": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "message"
          },
          "code": {
            "type": "integer",
            "example": "403"
          },
          "data": {
            "type": "object"
          }
        },
        "additionalProperties": false
      },
      "errorResponse500": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Internal server error"
          },
          "code": {
            "type": "integer",
            "example": "500"
          },
          "data": {
            "type": "object"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "authentication",
        "description": "JWT Authorization header using the apiKeyAuth."
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "All Auth module apis"
    },
    {
      "name": "Resource",
      "description": "All Resource module apis"
    }
  ]
}